Init commit
Finished the entire project on flip servers without uploading any code to github. Because of this, the inital commit will be the final commit. Sorry about the readability issues in the code. I kind of just made sure that the code worked and from there, didnt put too many comments in LOL. feel free to ask if you wanna know about a line of code.